Nesting Application components into a folder
