+++ client/src/components/Application/FundingCategoryPages/Health.js
@@ -1,20 +1,20 @@
 import React from "react";
-import { PreviousNext } from "../PreviousNext";
+import { PreviousNext } from "../../PreviousNext";
 import styled from "styled-components";
-import { TextInput } from "../FormComponents/TextInput";
+import { TextInput } from "../../FormComponents/TextInput";
 
-export const LivingExpenses = () => {
+export const Health = () => {
   return (
     <Container>
-      <h3>Living Expenses</h3>
+      <h3>Health</h3>
       <form>
-        {/* <label htmlFor="LIVING_EXPENSES">
+        {/* <label htmlFor="HEALTH">
           If applicable, please provide a detailed description of any resources
-          you wish to pursue funding for within the Living Expenses category. If
-          you are not requesting funds for this category, just enter N/A:
+          you wish to pursue funding for within the Health category. If you are
+          not requesting funds for this category, just enter N/A:
         </label>
         <br />
-        <textarea rows="8" id="LIVING_EXPENSES" name="LIVING_EXPENSES" />
+        <textarea rows="8" id="HEALTH" name="HEALTH" />
         <br /> */}
         <TextInput
           fieldName="amount"
@@ -22,7 +22,7 @@ export const LivingExpenses = () => {
           label="What is the total dollar amount being requested for this category?"
         />
       </form>
-      <PreviousNext prevLink="/funding" nextLink="/funding/insurance" />
+      <PreviousNext prevLink="/funding/programs" nextLink="/funding/travel" />
     </Container>
   );
 };
