+++ client/src/components/Application/FundingCategoryPages/Insurance.js
@@ -1,20 +1,20 @@
 import React from "react";
-import { PreviousNext } from "../PreviousNext";
+import { PreviousNext } from "../../PreviousNext";
 import styled from "styled-components";
-import { TextInput } from "../FormComponents/TextInput";
+import { TextInput } from "../../FormComponents/TextInput";
 
-export const Programs = () => {
+export const Insurance = () => {
   return (
     <Container>
-      <h3>Programs</h3>
+      <h3>Insurance</h3>
       <form>
-        {/* <label htmlFor="PROGRAMS">
+        {/* <label htmlFor="INSURANCE">
           If applicable, please provide a detailed description of any resources
-          you wish to pursue funding for within the Programs category. If you
+          you wish to pursue funding for within the Insurance category. If you
           are not requesting funds for this category, just enter N/A:
         </label>
         <br />
-        <textarea rows="8" id="PROGRAMS" name="PROGRAMS" />
+        <textarea rows="8" id="INSURANCE" name="INSURANCE" />
         <br /> */}
         <TextInput
           fieldName="amount"
@@ -23,8 +23,8 @@ export const Programs = () => {
         />
       </form>
       <PreviousNext
-        prevLink="/funding/winter-equipment"
-        nextLink="/funding/health"
+        prevLink="/funding/living-expenses"
+        nextLink="/funding/adaptive-equipment"
       />
     </Container>
   );
