+++ client/src/components/Application/FundingCategoryPages/Programs.js
@@ -1,27 +1,31 @@
 import React from "react";
-import { PreviousNext } from "../PreviousNext";
+import { PreviousNext } from "../../PreviousNext";
 import styled from "styled-components";
-import { TextInput } from "../FormComponents/TextInput";
+import { TextInput } from "../../FormComponents/TextInput";
 
-export const Travel = () => {
+export const Programs = () => {
   return (
     <Container>
-      <h3>Travel</h3>
+      <h3>Programs</h3>
       <form>
-        {/* <label htmlFor="TRAVEL">
+        {/* <label htmlFor="PROGRAMS">
           If applicable, please provide a detailed description of any resources
-          you wish to pursue funding for within the Travel category. If you are
-          not requesting funds for this category, just enter N/A:
+          you wish to pursue funding for within the Programs category. If you
+          are not requesting funds for this category, just enter N/A:
         </label>
         <br />
-        <textarea rows="8" id="TRAVEL" name="TRAVEL" /> */}
+        <textarea rows="8" id="PROGRAMS" name="PROGRAMS" />
+        <br /> */}
         <TextInput
           fieldName="amount"
           type="number"
           label="What is the total dollar amount being requested for this category?"
         />
       </form>
-      <PreviousNext prevLink="/funding/health" nextLink="/personal-info" />
+      <PreviousNext
+        prevLink="/funding/winter-equipment"
+        nextLink="/funding/health"
+      />
     </Container>
   );
 };
