+++ client/src/components/Application/FundingCategoryPages/Travel.js
@@ -1,28 +1,27 @@
 import React from "react";
-import { PreviousNext } from "../PreviousNext";
+import { PreviousNext } from "../../PreviousNext";
 import styled from "styled-components";
-import { TextInput } from "../FormComponents/TextInput";
+import { TextInput } from "../../FormComponents/TextInput";
 
-export const Health = () => {
+export const Travel = () => {
   return (
     <Container>
-      <h3>Health</h3>
+      <h3>Travel</h3>
       <form>
-        {/* <label htmlFor="HEALTH">
+        {/* <label htmlFor="TRAVEL">
           If applicable, please provide a detailed description of any resources
-          you wish to pursue funding for within the Health category. If you are
+          you wish to pursue funding for within the Travel category. If you are
           not requesting funds for this category, just enter N/A:
         </label>
         <br />
-        <textarea rows="8" id="HEALTH" name="HEALTH" />
-        <br /> */}
+        <textarea rows="8" id="TRAVEL" name="TRAVEL" /> */}
         <TextInput
           fieldName="amount"
           type="number"
           label="What is the total dollar amount being requested for this category?"
         />
       </form>
-      <PreviousNext prevLink="/funding/programs" nextLink="/funding/travel" />
+      <PreviousNext prevLink="/funding/health" nextLink="/personal-info" />
     </Container>
   );
 };
