+++ client/src/components/Application/FundingCategoryPages/index.js
@@ -1,0 +1,8 @@
+export { HealingNetwork } from "./HealingNetwork";
+export { LivingExpenses } from "./LivingExpenses";
+export { Insurance } from "./Insurance";
+export { AdaptiveEquipment } from "./AdaptiveEquipment";
+export { WinterEquipment } from "./WinterEquipment";
+export { Programs } from "./Programs";
+export { Health } from "./Health";
+export { Travel } from "./Travel";
