--- client/src/components/FundingCategoryPages/index.js
@@ -1,19 +1,0 @@
-import { HealingNetwork } from "./HealingNetwork";
-import { LivingExpenses } from "./LivingExpenses";
-import { Insurance } from "./Insurance";
-import { AdaptiveEquipment } from "./AdaptiveEquipment";
-import { WinterEquipment } from "./WinterEquipment";
-import { Programs } from "./Programs";
-import { Health } from "./Health";
-import { Travel } from "./Travel";
-
-export {
-  HealingNetwork,
-  LivingExpenses,
-  Insurance,
-  AdaptiveEquipment,
-  WinterEquipment,
-  Programs,
-  Health,
-  Travel
-};
